Add ability to retain git-tmp after builds to help future debugging efforts.

+++ cron.sh
@@ -1,18 +1,41 @@
 #!/bin/bash
 
 # ==
+# == Parse args
+# ==
+
+POSITIONAL=()
+while [ "$#" != '0' ]; do
+  case "$1" in
+    # useful for wanting to save git-tmp and observe commits/build artifacts
+    -n|--no-cleanup)
+      CLEANUP='no'
+      shift
+      ;;
+    *)
+      POSITIONAL+=("$1")
+      shift
+      ;;
+  esac
+done
+set -- "${POSITIONAL[@]}"
+
+# ==
 # == Setup
 # ==
 
+# export to expose to python script later on
 export RUN_DIR="$(pwd)"
 export GIT_TMP="$RUN_DIR/git-tmp"
 
 exit_routine() {
   local -r exit_status="${1}"
 
+  if [ "${CLEANUP:-yes}" = 'yes' ]; then
     if [ -d "${GIT_TMP}" ]; then
         rm -rf "${GIT_TMP}"
     fi
+  fi
 
   exit "${exit_status}"
 }
@@ -35,10 +58,12 @@ if [ "$ENV" = 'prod' ] || [ "$ENV" = 'test' ]; then
     git config --global user.name 'dbt-hubcap'
 fi
 
+# hubcap expects a fresh git-tmp, so deletion of git-tmp is forced before script is allowed to run in full
 if ! [ -d "${GIT_TMP}" ]; then
     mkdir "${GIT_TMP}"
 else
-    >&2 printf "Error: ${GIT_TMP} already exists. Deleting and exiting."
+    >&2 printf "Error: ${GIT_TMP} already exists. Deleting and exiting.\n"
+    unset CLEANUP
     exit 1
 fi
